chat

+++ src/main/java/com/example/ecommerce/realtime/service/chat/MessageService.java
@@ -13,7 +13,7 @@ public interface MessageService {
 
     Message getMessageById(Long id);
 
-    int getTotalUnreadMessageFromUserId(Long fromUserId);
+    Long getTotalUnreadMessageFromUserId(Long fromUserId);
 
     @Async
     void updateMessageStatusFromReadMessages(Long fromUserId, Long toUserId);
